ProjectFinalv1.1/requirements.txt
pytesseract
Pillow
googletrans==4.0.0-rc1
# You must also install Tesseract-OCR from https://github.com/tesseract-ocr/tesseract 
easyocr
torch 